requirements.txt
streamlit==1.37.1
pillow==10.4.0
qrcode==7.4.2
python-barcode[images]==0.15.1
reportlab==4.2.2